+++ lib/legion/extensions/llm/ledger/transport/transport.rb
@@ -2,13 +2,16 @@
 
 begin
   require 'legion/extensions/transport'
+  require 'legion/llm/transport/exchanges/audit'
+  require 'legion/llm/transport/exchanges/escalation'
+  require 'legion/llm/transport/exchanges/metering'
 rescue LoadError => _e
   nil
 end
 
 module Legion
   module Extensions
-    module LLM
+    module Llm
       module Ledger
         module Transport
           extend Legion::Extensions::Transport if Legion::Extensions.const_defined?(:Transport, false)
@@ -17,17 +20,17 @@ module Legion
             [
               {
                 from:        Legion::LLM::Transport::Exchanges::Metering,
-                to:          Legion::Extensions::LLM::Ledger::Transport::Queues::MeteringWrite,
+                to:          Legion::Extensions::Llm::Ledger::Transport::Queues::MeteringWrite,
                 routing_key: 'metering.#'
               },
               {
                 from:        Legion::LLM::Transport::Exchanges::Audit,
-                to:          Legion::Extensions::LLM::Ledger::Transport::Queues::AuditPrompts,
+                to:          Legion::Extensions::Llm::Ledger::Transport::Queues::AuditPrompts,
                 routing_key: 'audit.prompt.#'
               },
               {
                 from:        Legion::LLM::Transport::Exchanges::Audit,
-                to:          Legion::Extensions::LLM::Ledger::Transport::Queues::AuditTools,
+                to:          Legion::Extensions::Llm::Ledger::Transport::Queues::AuditTools,
                 routing_key: 'audit.tool.#'
               }
             ]